+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - remove **(Single)**
   - do not remove **label** from the front if it is not followed by a space
 
+* `artist`: keep the original artist separator in releases with a single track
+
 
 [0.16.0]: https://github.com/snejus/beetcamp/releases/tag/0.16.0